// guig/shared/types.ts
export interface BranchRef {
  name: string;
  fullName: string;
  sha: string;
  current: boolean;
  remote: boolean;
  upstream?: string;
  tracking?: string;
}

export interface Commit {
  sha: string;
  parents: readonly string[];
  author: string;
  authorEmail: string;
  authoredAt: string;
  committer: string;
  committerEmail: string;
  committedAt: string;
  subject: string;
  body?: string;
  decorations: readonly string[];
}